Enum mechylang::evaluator::objects::Object

pub enum Object {

    Integer(i64),
    Float(f64),
    Boolean(bool),
    Unit,
    ReturnValue(Box<Object>),
    Function(Function),
    String(Rc<str>),
    BuiltinFunction(BuiltinFunction),
    Array(Vec<Object>),
    RangeFrom(Rc<Object>),
    RangeTo(Rc<Object>),
    RangeToInclusive(Rc<Object>),
    Range(Rc<Object>, Rc<Object>),
    RangeInclusive(Rc<Object>, Rc<Object>),
    RangeFull,
    Iterator(IteratorObject),
    Break(Option<Box<Object>>),
    Continue,
    Method(Method),
    Reference(Reference),
    Struct(HashMap<String, Object>),

}

Variants

Integer(i64)

Integer

An integer is a whole number. It can be positive or negative.

let x = 5;
assert_eq(x, 5);

Max and Min

Integers have a maximum and minimum value. If you try to assign a value that is too large or too small, you will get an error.

Max: 263 - 1 = 9,223,372,036,854,775,807 Min: -263 = -9,223,372,036,854,775,808

Float(f64)

Float

A float is a decimal number. It can be positive or negative.

let x = 5.5;
assert_eq(x, 5.5);

Max and Min

Floats have a maximum and minimum value. If you try to assign a value that is too large or too small, you will get an error.

Max: 1.7976931348623157308 Min: 2.2250738585072014-308

Boolean(bool)

Unit

ReturnValue(Box<Object>)

Function(Function)

String(Rc<str>)

BuiltinFunction(BuiltinFunction)

Array(Vec<Object>)

Array

An array is a mutable, ordered list of items. It can contain any type of object.

Usage

To create an array, use the following syntax:

let arr = [1, 2, 3];
assert_eq(arr, [1, 2, 3]);

Indexing

To get an item from an array, use the following syntax:

let arr = [1, 2, 3];
assert_eq(arr[0], 1);
assert_eq(arr[1], 2);
assert_eq(arr[2], 3);

Methods

Arrays have many methods, take a look at the array methods module for more information.

RangeFrom(Rc<Object>)

RangeTo(Rc<Object>)

RangeToInclusive(Rc<Object>)

Range(Rc<Object>, Rc<Object>)

RangeInclusive(Rc<Object>, Rc<Object>)

RangeFull

Iterator(IteratorObject)

Break(Option<Box<Object>>)

Continue

Method(Method)

Reference(Reference)

Struct(HashMap<String, Object>)