mechylang::evaluator::runtime

Module builtins

Builtins for mechylang

print

The print function can be used to print to stdout. It prints the string representation of the arguments passed to it, separated by spaces.

print("Hello", "World!"); // prints "Hello World!"

println

The println function can be used to print to stdout with a newline at the end. It prints the string representation of the arguments passed to it, separated by spaces.

println("Hello", "World!"); // prints "Hello World!\n"

assert

The assert function can be used to assert that a boolean is true.

assert(true); // ok

assert(false); // panics

assert_eq

The assert_eq function can be used to assert that all arguments are equal.

assert_eq(1, 1, 1); // ok

assert_eq(1, 2, 1); // panics

input_read_all

The input_read_all function can be used to read all input from stdin. It keeps reading until EOF is reached.

let input = input_read_all();
assert_eq(input, "Hello\nWorld\n"); // ok

input_read_line

The input_read_line function can be used to read a single line from stdin. It reads until a newline is reached.

let input = input_read_line();
assert_eq(input, "Hello\n"); // ok

parse_int

The parse_int function can be used to parse a string into an integer.

let num = parse_int("123");
assert_eq(num, 123); // ok

It throws an error if the string is not a valid integer.

let num = parse_int("123a");

parse_float

The parse_float function can be used to parse a string into a float.

let num = parse_float("123.45");
assert_eq(num, 123.45); // ok

It throws an error if the string is not a valid float.

let num = parse_float("123.45a");

Structs

BUILTINS

A list of all the builtin functions in mechylang

BuiltinFunction

A builtin function that can be called from mechylang

Enums

BuiltinError