mechylang::evaluator::runtime

Module environment

Environment

The environment is a data structure that stores variables and their values for mechylang programs. The environment is a tree of scopes, where each scope has access to the variables in connected outer scopes. The environment is used to store variables and their values during the evaluation of a mechylang program.

Scope

A scope is a mapping of variable names to their values. A scope can be thought of as a dictionary or a hashmap. A scope can be empty, in which case it has no variables and no values. A scope can also be non-empty, in which case it has variables and values. A scope can also be nested inside another scope. A scope that is nested inside another scope is called an enclosed scope. An enclosed scope has access to the variables and values in the outer scope. A scope that is not nested inside another scope is called a global scope. A global scope has no access to any other scope.

Global Scope {
   x: 1, y: 2, z: 3
   // has access to x, y, z
   Enclosed Scope {
        foo: 4, bar: 5
        // has access to x, y, z and foo, bar
        Enclosed Scope {
            baz: 6
            // has access to x, y, z, foo, bar and baz
        }
         
        Enclosed Scope {
            qux: 7
            // has access to x, y, z, foo, bar and qux
        }
   }
   Enclosed Scope {
        quux: 8
        // has access to x, y, z and quux
   }
}

Environment

An environment is a representation of a scope, which also has access to any outer scopes.

`mechylang uses one global heap to store all objects. Variables are stored per scope. This means that multiple scopes can have access to the same object on the heap, but each scope can only access the variables it has access to.

Shadowing

A scope can have a variable with the same name as a variable in an outer scope. In this case, the variable in the inner scope shadows the variable in the outer scope. This means that the variable in the outer scope is not accessible from the inner scope. Only the variable in the inner scope is accessible from the inner scope. This is called shadowing.

let x = 1;
assert_eq(x, 1);
{ // This is a new scope
    assert_eq(x, 1); // x is accessible from this scope
    let x = 2;
    assert_eq(x, 2);
} // The scope ends here
// In the outer scope, x is still 1
assert_eq(x, 1);

How does this work? When a variable is set in a scope, it is simply added to, or overwritten in the current scope. Any potential outer scopes are ignored.

Updating

A scope can also update a variable in an outer scope. This means that the variable in the outer scope is updated to the new value. This is called updating.

let x = 1;
assert_eq(x, 1);
{ // This is a new scope
    assert_eq(x, 1); // x is accessible from this scope
    x = 2;
    assert_eq(x, 2);
    { // This is a new scope
        assert_eq(x, 2); // x is accessible from this scope
        x = 3;
        assert_eq(x, 3);
    } // The scope ends here
    // In the outer scope, x is now 3
    assert_eq(x, 3);
} // The scope ends here
// In the outer scope, x is now 3
assert_eq(x, 3);

How does this work? When a variable is updated in a scope, the variable is first searched for in the current scope. If the variable is found in the current scope, then the variable is updated in the current scope. If the variable is not found in the current scope, then the variable is searched for in the outer scopes recursively. The value is updated in the first scope where the variable is found. If the variable is not found in any scope, then an error is returned.

For more info on updating see Environment::update

Example

use mechylang::Environment;
use mechylang::Object;

let mut env = Environment::new();
env.set("x", Object::Integer(1));
assert_eq!(env.get("x"), Some(Object::Integer(1)));

// The enclosed env also has access to the value
let mut enclosed_env = Environment::new_enclosed(&env);
assert_eq!(enclosed_env.get("x"), Some(Object::Integer(1)));

// The value can be set in the enclosed environment
enclosed_env.set("x", Object::Integer(2));
// Now this new value for x shadows the old value for x
assert_eq!(enclosed_env.get("x"), Some(Object::Integer(2)));
assert_eq!(env.get("x"), Some(Object::Integer(1)));

// If we want to update the first value found (by searching outwards from the current scope)
// we can use Environment::update
env.set("y", Object::Integer(1)); // Set a new variable as `x` is currently shadowed
assert_eq!(env.get("y"), Some(Object::Integer(1)));
assert_eq!(enclosed_env.get("y"), Some(Object::Integer(1)));

// Update the value of `y` in the enclosed environment
enclosed_env.update("y", Object::Integer(2)).unwrap();
// Now the value of `y` is updated in the outer environment
// which the enclosed environment has access to
assert_eq!(env.get("y"), Some(Object::Integer(2)));
assert_eq!(enclosed_env.get("y"), Some(Object::Integer(2)));

For more info on Environment::set, Environment::update, Environment::get and Environment::new_enclosed see their respective documentation.

Structs

Environment

An environment is a representation of a scope, which also has access to any outer scopes.

HeapObject

InnerEnvironment