mechylang::parser

Module expressions

Parsing Expressions

This module contains the parsing logic for expressions.

Expressions

An expression is a combination of values, variables, operators, and functions that are interpreted to produce another value. For example, 1 + 2 is an expression that evaluates to 3. Expressions can be as simple as a single value, or as complex as a function call with multiple arguments.

This file contains the following:

Expression enum

All expression variants

implementations for Display for all expression variants

implementations for the parse functions for all expression variants for the parser

Structs

ArrayLiteral

BlockExpression

Represents a block of code in Mechylang, which is a sequence of statements enclosed in curly braces.

BooleanLiteral

CallExpression

FloatLiteral

ForExpression

FunctionLiteral

Identifier

IfExpression

IndexExpression

InfixExpression

IntegerLiteral

MemberExpression

PrefixExpression

RangeExpression

RangeFromExpression

RangeFullExpression

RangeToExpression

StringLiteral

StructLiteral

WhileExpression

Enums

Expression

Represents an expression in Mechylang.

InfixOperator

Precedence

PrefixOperator

Traits

ExpressionSpanExt